tests/wpt/tests/fenced-frame/revoke-nested-fenced-frame-navigation.https.html

<!DOCTYPE html>
<title>Test that window.fence.disableUntrustedNetwork disables
  embedder-initiated navigation of nested fenced frames.</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="resources/utils.js"></script>

<body>
<script>

// Run a test with a fenced frame nested in a fenced frame.
// If `should_disable_network` is true, window.fence.disableUntrustedNetwork
// will be called before creating the nested fenced frame.
async function ff_ff_test(t, should_disable_network, should_succeed) {
  const fencedframe = await attachFencedFrameContext();
  const navigation_promise =
      fencedframe.execute(async (should_disable_network) => {
        if (should_disable_network) {
          await window.fence.disableUntrustedNetwork();
        }
        const nested_fenced_frame = await attachFencedFrameContext();
        return nested_fenced_frame.execute(() => { return 'nav success'; }); },
      [should_disable_network]);
  if (should_succeed) {
    const result = await navigation_promise;
    assert_equals(result, 'nav success');
  } else {
    const result = await Promise.race([
        navigation_promise,
        new Promise((resolve, reject) => t.step_timeout(
          () => resolve('timeout'), 2000))]);
    assert_equals(result, 'timeout');
  }
}

promise_test(async(t) => {
  await ff_ff_test(t, /*should_disable_network=*/false,
                      /*should_succeed=*/true);
}, 'FF->FF navigation works');

promise_test(async(t) => {
  await ff_ff_test(t, /*should_disable_network=*/true,
                      /*should_succeed=*/false);
}, 'window.fence.disableUntrustedNetwork disables FF->FF navigation');

</script>
</body>
